Moved SideBar.vue into components
